% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filedf}
\alias{filedf}
\title{Filedf}
\format{
A data frame containing 2 columns - document slug, and string copy of
the OCR output for the document
}
\usage{
filedf
}
\description{
The other two datasets are the OCR outputs for the pdf files.  Both are data
frames containing document slugs and string outputs, to be used by the document
search function.  The filedf contains just this.
}
\keyword{dataset}
