% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fileids}
\alias{fileids}
\title{File IDs}
\format{
A data frame containing 6 columns, including document name (and thus slug),
Clowder id, size, and more.
}
\usage{
fileids
}
\description{
The first is the fileids dataset.  This is used by the document pulling
functionality, to map document slugs identified to Clowder ids, to allow for
pulling the pdf from Clowder.
}
\keyword{dataset}
