% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pagedf}
\alias{pagedf}
\title{Pagedf}
\format{
A data frame containing 3 columns - document slug, page number, and
string copy of the OCR output for the document
}
\usage{
pagedf
}
\description{
OCR outputs for the pdf files.  Contains a third column for page number, with
string outputs split by page.
}
\keyword{dataset}
